+++ test/src/Dropdown/Dropdown.spec.js
@@ -344,30 +344,6 @@ describe('Dropdown[renderItem]', () => {
   })
 })
 
-describe('Dropdown[close]', () => {
-  test('should close', () => {
-    jest.useFakeTimers()
-    const wrapper = mount(<Dropdown absolute data={data} placeholder="Dropdown" />)
-    wrapper.find('button').simulate('click')
-    expect(
-      wrapper
-        .find('ShineoutDropdown')
-        .first()
-        .instance().state.show
-    ).toBe(true)
-    const click = new UIEvent('click')
-    click.initUIEvent('click')
-    document.dispatchEvent(click)
-    jest.runAllTimers()
-    expect(
-      wrapper
-        .find('ShineoutDropdown')
-        .first()
-        .instance().state.show
-    ).toBe(false)
-  })
-})
-
 describe('Dropdown[trigger]', () => {
   test('should set trigger hover', () => {
     jest.useFakeTimers()
